# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Servanda Office API
  version: 1.0.0
  description: |
    Multi-Tenant LegalTech Platform API for contract creation and management.

    **Core Flow:** Template selection -> Guided Q&A Interview -> Conflict-free Contract -> DOCX/ODT Export.

    **Architecture Decisions:**
    - ADR-001: Shared DB with PostgreSQL Row-Level Security for tenant isolation
    - ADR-002: Immutable versioning for Clauses/Templates; ContractInstance pins versions
    - ADR-003: Async export via pgboss job queue with S3 storage
    - ADR-004: Server-side ODT conversion via LibreOffice headless

    **Authentication:**
    - Production: Bearer JWT (OIDC/SAML)
    - Development: Header-based tenant context (x-tenant-id, x-user-id, x-user-role)

    **Pagination:** All list endpoints support `page` and `pageSize` query parameters.
    Default page size: 20, maximum: 100.

    **Status Workflow (Clause/Template Versions):**
    `draft` -> `review` -> `approved` -> `published` -> `deprecated`
    (review -> draft and approved -> draft are also valid transitions)
  contact:
    name: Team 01 - Product Architecture
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server (v1 canonical prefix)
  - url: http://localhost:3000/api
    description: Legacy route prefix (backward-compatible)
security:
  - BearerAuth: []
  - DevHeaders: []
...
paths:
...
  /identity/me:
    get:
      tags: [Identity]
      summary: Get current user info
      description: Returns the currently authenticated user's profile.
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /identity/users:
    get:
      tags: [Identity]
      summary: List users in tenant
      description: Returns a paginated list of all users belonging to the current tenant.
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /identity/users/invite:
    post:
      tags: [Identity]
      summary: Invite a new user
      description: |
        Creates a new user with status `invited`. Admin role required.
        Email must be unique within the tenant.
      operationId: inviteUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteUserRequest'
      responses:
        '201':
          description: User invited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /identity/users/{id}:
    get:
      tags: [Identity]
      summary: Get a single user
      description: Returns a user by ID within the current tenant.
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Identity]
      summary: Update a user
      description: Updates user role and/or display name. Admin role required.
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Identity]
      summary: Delete a user
      description: |
        Permanently deletes a user. Admin role required.
        Cannot delete yourself.
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: User deleted successfully
        '400':
          description: Cannot delete yourself
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /identity/users/{id}/activate:
    post:
      tags: [Identity]
      summary: Activate a user
      description: |
        Transitions user status to `active`. Admin role required.
        Fails if user is already active.
      operationId: activateUser
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User activated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: User is already active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /identity/users/{id}/deactivate:
    post:
      tags: [Identity]
      summary: Deactivate a user
      description: |
        Transitions user status to `inactive`. Admin role required.
        Cannot deactivate yourself. Fails if user is already inactive.
      operationId: deactivateUser
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User deactivated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Cannot deactivate yourself
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: User is already inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # =========================================================================
  # IDENTITY - Audit Logs
  # =========================================================================
  /identity/audit-logs:
    get:
      tags: [Identity]
      summary: Query audit logs
      description: |
        Returns paginated audit events for the current tenant. Admin role required.
        Supports filtering by action, object type, object ID, and time range.
      operationId: queryAuditLogs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: action
          in: query
          description: Filter by audit action (e.g. `clause.create`, `user.invite`)
          schema:
            type: string
        - name: objectType
          in: query
          description: Filter by object type (e.g. `clause`, `user`, `export_job`)
          schema:
            type: string
        - name: objectId
          in: query
          description: Filter by specific object ID
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          description: Filter events from this timestamp (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Filter events until this timestamp (ISO 8601)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Paginated audit log entries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AuditEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # =========================================================================
  # CONTENT - Clauses
  # =========================================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Production authentication via OIDC/SAML JWT token.
        Include tenant claim in the JWT payload.
    DevHeaders:
      type: apiKey
      in: header
      name: x-tenant-id
      description: |
        Development-mode authentication via headers.
        Requires all three headers:
        - `x-tenant-id` (UUID) - Tenant identifier
        - `x-user-id` (UUID) - User identifier
        - `x-user-role` (admin|editor|user) - User role

  # =========================================================================
  # Parameters
  # =========================================================================
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      description: Page number (1-based, default 1)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Number of items per page (default 20, max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # =========================================================================
  # Reusable Responses
  # =========================================================================
  responses:
    BadRequest:
      description: Invalid request body or parameters (Zod validation failed)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Insufficient permissions (role check failed)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  # =========================================================================
  # Schemas
  # =========================================================================
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: NOT_FOUND
        message:
          type: string
          description: Human-readable error message
          example: Resource not found
        details:
          type: object
          additionalProperties: true
          description: Additional error context
    PaginatedResponse:
      type: object
      required: [data, total, page, pageSize, hasMore]
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
          description: Total number of items across all pages
          example: 42
        page:
          type: integer
          description: Current page number
          example: 1
        pageSize:
          type: integer
          description: Items per page
          example: 20
        hasMore:
          type: boolean
          description: Whether more pages exist
          example: true

    # --- Identity ---
    User:
      type: object
      required: [id, email, displayName, role, status, mfaEnabled, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          type: string
          enum: [admin, editor, user]
        status:
          type: string
          enum: [invited, active, inactive]
        mfaEnabled:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    InviteUserRequest:
      type: object
      required: [email, displayName, role]
      properties:
        email:
          type: string
          format: email
          description: Must be unique within the tenant
        displayName:
          type: string
          minLength: 1
          maxLength: 255
        role:
          type: string
          enum: [admin, editor, user]
    UpdateUserRequest:
      type: object
      properties:
        role:
          type: string
          enum: [admin, editor, user]
        displayName:
          type: string
          minLength: 1
          maxLength: 255
    AuditEvent:
      type: object
      required: [id, tenantId, userId, action, objectType, objectId, timestamp]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        action:
          type: string
          description: Audit action identifier
          example: clause.create
        objectType:
          type: string
          description: Type of the affected object
          example: clause
        objectId:
          type: string
          format: uuid
        details:
          type: object
          additionalProperties: true
        ip:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time

    # --- Content: Clauses ---
...